{{cookiecutter.project_slug}}/requirements.txt
asgiref==3.8.1
attrs==25.3.0
dj-database-url==3.0.0
Django==5.2.3
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
django-modeltranslation>=0.18.0
django-otp>=1.2.2
qrcode>=7.3.1
Pillow>=9.0.0
pyotp>=2.6.0
boto3>=1.20.0
django-storages>=1.12.0
